Fix dependencies to the new snapshots project
