+++ subprojects/language-java/language-java.gradle
@@ -1,3 +1,4 @@
+import org.gradle.gradlebuild.testing.integrationtests.cleanup.WhenNotEmpty
 import org.gradle.gradlebuild.unittestandcompile.ModuleType
 
 plugins {
@@ -13,6 +14,7 @@ dependencies {
     compile project(":core")
     compile project(":platformJvm")
     compile project(":languageJvm")
+    implementation project(":snapshots")
 
     // TODO - get rid of this cycle
     integTestRuntime project(':plugins')
@@ -36,7 +38,6 @@ classycle {
                        'org/gradle/external/javadoc/**']
 }
 
-import org.gradle.gradlebuild.testing.integrationtests.cleanup.WhenNotEmpty
 testFilesCleanup {
     policy = WhenNotEmpty.REPORT
 }
