+++ subprojects/plugins/plugins.gradle
@@ -1,3 +1,4 @@
+import org.gradle.gradlebuild.testing.integrationtests.cleanup.WhenNotEmpty
 import org.gradle.gradlebuild.unittestandcompile.ModuleType
 
 /*
@@ -29,6 +30,7 @@ dependencies {
     compile project(':languageGroovy')
     compile project(':diagnostics')
     compile project(':testingJvm')
+    compile project(':snapshots')
 
     compile libraries.ant.coordinates
     compile libraries.asm.coordinates
@@ -71,7 +73,6 @@ def wrapperJar = tasks.register("wrapperJar", Copy) {
 }
 sourceSets.main.output.dir wrapperJarDir, builtBy: wrapperJar
 
-import org.gradle.gradlebuild.testing.integrationtests.cleanup.WhenNotEmpty
 testFilesCleanup {
     policy = WhenNotEmpty.REPORT
 }
